replace scroll views with wheelViews.
the generic view got replaced with wheelview
as part of the app facelift, because it didn't gave
the user any indication it got pressed and dragged.
now user can see a black squere to indicate movement.

the WheelView was introduced using an open-source
custom view which implements this behaviour.
the director 'libs/wheel.jar' contains it.
the code is availiable at:
http://code.google.com/p/android-wheel/

Signed-off-by: Ohad Moisi <ohad.moisi@tandemg.com>