ScratchpadActivity: add generic keyboard events.
add support to the 'v' command, which will receive int or chars
to be passed directly to the keyboard driver on the DK40
for example, inorder to change the volume up\down we will use:

"v 93" will create a volume up event.
"v 94" will trigger volume down.

specific values can be found in a map located in the toolbox

Signed-off-by: Ohad Moisi <ohad.moisi@tandemg.com>